+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+## [v.0.5.2]
+
+- Add [#11]: Quality Summary and Preferred data sources in verification
+- Add [#9]: Additional information how to install in README.md
 - Add [#8]: Retry verification if any error happens in the process.
 - Add [#7]: Add EditDistance field to verification output.
 - Add [#6]: Add 'NoMatch' value to verification 'MatchType'.
@@ -31,9 +35,12 @@
 
 This document follows [changelog guidelines]
 
+[v0.5.2]: https://github.com/gnames/gnfinder/compare/v0.5.1...v0.5.2
 [v0.5.1]: https://github.com/gnames/gnfinder/compare/v0.5.0...v0.5.1
 [v0.5.0]: https://github.com/gnames/gnfinder/tree/v0.5.0
 
+[#11]: https://github.com/gnames/gnfinder/issues/11
+[#9]: https://github.com/gnames/gnfinder/issues/9
 [#8]: https://github.com/gnames/gnfinder/issues/8
 [#7]: https://github.com/gnames/gnfinder/issues/7
 [#6]: https://github.com/gnames/gnfinder/issues/6